fixing auth route page
